refactor: removed deprecated dataset row.
