+++ frontend/src/components/Collection/components/CollectionDatasetsGrid/components/DatasetsGrid/index.tsx
@@ -1,9 +1,3 @@
-/* TODO(cc) review (#1718). */
-/* Copied from src/components/Collections/components/Grid/components/DatasetsGrid and modified as an intermediate
-   upgrade to the collection datasets table while keeping the existing core datasets grid outside of the filter feature
-   flag untouched. Once filter feature flag is removed, the existing core datasets grid can be deleted and replaced
-   with this version. Ideally collection datasets table should be moved to react-table but requires changes to how
-   dataset information (statuses etc) is calculated and rolled out. */
 import { FC } from "react";
 import { UseMutateAsyncFunction } from "react-query";
 import { Collection, Dataset, VISIBILITY_TYPE } from "src/common/entities";
