+++ frontend/src/components/Collection/components/CollectionDatasetsGrid/components/Row/DatsetRow/components/MoreDropdown/index.tsx
@@ -1,7 +1,7 @@
 import { Position } from "@blueprintjs/core";
 import { useMemo } from "react";
 import RawMoreDropdown from "src/components/common/MoreDropdown";
-import { Props as ChooserProps } from "src/components/DropboxChooser/index";
+import { Props as ChooserProps } from "src/components/DropboxChooser";
 import Menu from "./components/Menu";
 
 interface Props {
