+++ frontend/src/components/Collection/components/CollectionDatasetsGrid/components/Row/DatsetRow/index.tsx
@@ -1,8 +1,3 @@
-/* Copied from src/components/Collections/components/Grid/components/Row/DatasetRow and modified as an intermediate
-   upgrade to the collection datasets table while keeping the existing core datasets grid outside of the filter feature
-   flag untouched. Once filter feature flag is removed, the existing core datasets grid can be deleted and replaced
-   with this version.Ideally collection datasets table should be moved to react-table but requires changes to how
-   dataset information (statuses etc) is calculated and rolled out. */
 import { Intent, Tooltip } from "@blueprintjs/core";
 import loadable from "@loadable/component";
 import { FC } from "react";
@@ -19,10 +14,8 @@ import {
   VISIBILITY_TYPE,
 } from "src/common/entities";
 import { useDatasetStatus } from "src/common/queries/datasets";
-import DownloadDataset from "src/components/Collections/components/Dataset/components/DownloadDataset";
-import { aggregateDatasetsMetadata } from "src/components/Collections/components/Grid/common/utils";
-import MoreDropdown from "src/components/Collections/components/Grid/components/Row/DatasetRow/components/MoreDropdown";
-import RevisionStatusTag from "src/components/Collections/components/Grid/components/Row/DatasetRow/components/RevisionStatusTag";
+import MoreDropdown from "src/components/Collection/components/CollectionDatasetsGrid/components/Row/DatsetRow/components/MoreDropdown";
+import RevisionStatusTag from "src/components/Collection/components/CollectionDatasetsGrid/components/Row/DatsetRow/components/RevisionStatusTag";
 import {
   checkIfCancelled,
   checkIfFailed,
@@ -35,7 +28,9 @@ import {
   useCheckCollectionPopulated,
   useConversionProgress,
   useUploadProgress,
-} from "src/components/Collections/components/Grid/components/Row/DatasetRow/utils";
+} from "src/components/Collection/components/CollectionDatasetsGrid/components/Row/DatsetRow/utils";
+import DownloadDataset from "src/components/Collections/components/Dataset/components/DownloadDataset";
+import { aggregateDatasetsMetadata } from "src/components/Collections/components/Grid/common/utils";
 import { OVER_MAX_CELL_COUNT_TOOLTIP } from "src/components/common/Grid/common/constants";
 import { checkIsOverMaxCellCount } from "src/components/common/Grid/common/utils";
 import ActionButton from "src/components/common/Grid/components/ActionButton";
@@ -53,14 +48,14 @@ import { Props as ChooserProps } from "src/components/DropboxChooser/index";
 const AsyncTooltip = loadable(
   () =>
     /*webpackChunkName: 'Grid/Row/DatasetRow/Tooltip' */ import(
-      "src/components/Collections/components/Grid/components/Row/DatasetRow/components/Tooltip"
+      "src/components/Collection/components/CollectionDatasetsGrid/components/Row/DatsetRow/components/Tooltip"
     )
 );
 
 const AsyncUploadStatus = loadable(
   () =>
     /*webpackChunkName: 'Grid/Row/DatasetRow/UploadStatus' */ import(
-      "src/components/Collections/components/Grid/components/Row/DatasetRow/components/UploadStatus"
+      "src/components/Collection/components/CollectionDatasetsGrid/components/Row/DatsetRow/components/UploadStatus"
     )
 );
